+++ bin/parser.js
@@ -11,15 +11,12 @@ const fs = require('fs');
 const Y2R = require('../lib/rml-generator.js');
 const Y2R2 = require('../lib/r2rml-generator.js');
 const { canonicalize } = require('../lib/tools');
+const getNamespace = require('../lib/namespaces');
 const pkginfo = require('pkginfo');
 const N3 = require('n3');
-const namespaces = require('prefix-ns').asMap();
-namespaces['rml'] = 'http://semweb.mmlab.be/ns/rml#'; // this one is the only official one for now.
 const watch = require('../lib/watcher.js');
 const glob = require('glob');
 const Logger = require('../lib/logger');
-
-namespaces.ql = 'http://semweb.mmlab.be/ns/ql#';
 
 pkginfo(module, 'version');
 
@@ -82,15 +79,15 @@ if (!options.input) {
       let triples;
 
       let prefixes = {
-        rr: namespaces.rr,
-        rdf: namespaces.rdf,
-        rdfs: namespaces.rdfs,
-        fnml: "http://semweb.mmlab.be/ns/fnml#",
-        fno: "https://w3id.org/function/ontology#",
-        d2rq: "http://www.wiwiss.fu-berlin.de/suhl/bizer/D2RQ/0.1#",
-        void: "http://rdfs.org/ns/void#",
-        dc: "http://purl.org/dc/terms/",
-        foaf: "http://xmlns.com/foaf/0.1/"
+        rr: getNamespace('rr'),
+        rdf: getNamespace('rdf'),
+        rdfs: getNamespace('rdfs'),
+        fnml: getNamespace('fnml'),
+        fno: getNamespace('fno'),
+        d2rq: getNamespace('d2rq'),
+        void: getNamespace('void'),
+        dc: getNamespace('dc'),
+        foaf: getNamespace('foaf')
       };
 
       const externalReferences = {};
@@ -106,8 +103,8 @@ if (!options.input) {
         const y2r = new Y2R({ class: !!options.class, externalReferences, includeMetadata });
         triples = y2r.convert(inputData);
 
-        prefixes.rml = namespaces.rml;
-        prefixes.ql = namespaces.ql;
+        prefixes.rml = getNamespace('rml');
+        prefixes.ql = getNamespace('ql');
         prefixes[''] = y2r.getBaseIRI();
         prefixes = Object.assign({}, prefixes, y2r.getPrefixes());
       } else {